backend/requirements.txt
asgiref==3.6.0
black==23.1.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.1
Deprecated==1.2.13
distlib==0.3.6
Django==4.1.5
django-cors-headers==3.13.0
django-mysql==4.9.0
django-oauth-toolkit==2.2.0
django-webpack-loader==1.8.0
djangorestframework==3.14.0
filelock==3.9.0
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.58.0
httplib2==0.21.0
idna==3.4
jwcrypto==1.4.2
mypy-extensions==1.0.0
mysqlclient==2.1.1
oauth2client==4.1.3
oauthlib==3.2.2
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==3.0.0
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14
virtualenv==20.19.0
whitenoise==6.3.0
wrapt==1.14.1